requirements.txt
elasticsearch==7.0.5
PyYAML==5.1.2
urllib3==1.25.6
